+++ package.json
@@ -122,6 +122,7 @@
     "isomorphic-unfetch": "^2.0.0",
     "load-script": "^1.0.0",
     "lodash.debounce": "^4.0.8",
+    "lodash.get": "^4.0.8",
     "lru-cache": "^4.1.1",
     "markdown-in-js": "^1.1.3",
     "mdast-react-render": "^1.1.0",
